# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EffectiveMobile - TZ",
        "description": "Техническое задание для EffectiveMobile",
        "contact": {
            "email": "aravion1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Получить все задачи",
                "operationId": "get_tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskResource"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Создание задачи",
                "operationId": "create_task",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v1/tasks/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Получить задачу",
                "operationId": "get_task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Изменить задачу",
                "operationId": "update_task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "удалить задачу",
                "operationId": "delete_task",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateOrUpdateRequest": {
                "description": "Модель данных для создания или обновления задач",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "string"
                    },
                    "description": {
                        "type": "string",
                        "example": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "NEW",
                            "COMPLETED",
                            "CLOSED"
                        ],
                        "example": "NEW"
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "title": "TaskResource",
                "description": "Задача",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "string"
                    },
                    "description": {
                        "type": "string",
                        "example": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "NEW",
                            "COMPLETED",
                            "CLOSED"
                        ],
                        "example": "NEW"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Task",
            "description": "Методы управления задачами"
        }
    ]
}